requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
aiosqlite>=0.20.0
mcp>=1.0.0
numpy>=2.0.0
scipy>=1.13.0
matplotlib>=3.9.0
shapely>=2.0.0
geopandas>=1.0.0
python-docx>=1.1.0
reportlab>=4.2.0
httpx>=0.27.0
